// web/splive/src/app/hpft/app/models/api/patient-data-models.ts
export class PatientUpdateRequest {
    patientid: number;
    patientregno: string;
    fname: string;
    lname: string;
    mobno: string;
    age: string;
    bloodgrp: string;
    gender: string;
}